+++ Civi/DataProcessor/DataFlow/CombinedDataFlow/CombinedSqlDataFlow.php
@@ -66,7 +66,9 @@ class CombinedSqlDataFlow extends SqlDataFlow implements MultipleSourceDataFlows
    * @return string
    */
   public function getTableStatement() {
-    return "`{$this->primary_table}` `{$this->primary_table_alias}`";
+    $sourceDataFlowDescription = reset($this->sourceDataFlowDescriptions);
+    $dataFlow = $sourceDataFlowDescription->getDataFlow();
+    return $dataFlow->getTableStatement();
   }
 
   /**
@@ -211,6 +213,25 @@ class CombinedSqlDataFlow extends SqlDataFlow implements MultipleSourceDataFlows
    */
   public function getPrimaryTableAlias() {
     return $this->primary_table_alias;
+  }
+
+  /**
+   * @param \Civi\DataProcessor\DataFlow\SqlDataFlow\WhereClauseInterface $clause
+   *
+   * @return \Civi\DataProcessor\DataFlow\SqlDataFlow
+   */
+  public function removeWhereClause(SqlDataFlow\WhereClauseInterface $clause) {
+    foreach($this->whereClauses as  $i => $c) {
+      if ($c->getWhereClause() == $clause->getWhereClause()) {
+        unset($this->whereClauses[$i]);
+      }
+    }
+    foreach($this->sourceDataFlowDescriptions as $sourceDataFlowDescription) {
+      if ($sourceDataFlowDescription->getDataFlow() instanceof SqlDataFlow) {
+        $sourceDataFlowDescription->getDataFlow()->removeWhereClause($clause);
+      }
+    }
+    return $this;
   }