+++ Civi/DataProcessor/FilterHandler/ContactInGroupFilter.php
@@ -154,10 +154,12 @@ class ContactInGroupFilter extends AbstractFieldFilterHandler {
    * Add the elements to the filter form.
    *
    * @param \CRM_Core_Form $form
+   * @param array $defaultFilterValue
+   *
    * @return array
    *   Return variables belonging to this filter.
    */
-  public function addToFilterForm(\CRM_Core_Form $form) {
+  public function addToFilterForm(\CRM_Core_Form $form, $defaultFilterValue) {
     $fieldSpec = $this->getFieldSpecification();
     $operations = $this->getOperatorOptions($fieldSpec);
 
@@ -182,6 +184,7 @@ class ContactInGroupFilter extends AbstractFieldFilterHandler {
     ));
 
     $filter['type'] = $fieldSpec->type;
+    $filter['alias'] = $fieldSpec->alias;
     $filter['title'] = $title;
 
     return $filter;