+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Master version (in development)
+
+* Respect selected permissions for outputs 
+* Allow to specify "Is Empty" for various filters.
+* Allow to limit ContactFilter to only show contacts from specific groups.
+* it is now also possible to output a data processor on CiviCRMs dahsboard.
+* Added field outputs for simple calculations (substract and total).
+
+# Version 1.0.2
+
+* Fixed bug #11 (Fatal error clone on non object)
+
 # Version 1.0.1
 
 Initial release.